do MACs using EVP_MD_CTX with openssl >= 1.0.0
Using EVP_MD_CTX for MAC calculation allows to use other algorithms
than HMAC. In this particular case it allows the GOST-MAC to be used.
GOST-MAC uses a 256 bit key and produces a 32 bit signature.

Unfortunately OpenSSL has no API for querying a MAC's key length, so
the key length is returned statically in the case of GOST-MAC.

Signed-off-by: Heiko Hund <heiko.hund@sophos.com>

Ported to 2.4.6 version by Mikhail Efremov.